# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Management API",
        "contact": {
            "email": "info@tasks.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8006/api",
            "description": "Development Server"
        }
    ],
    "paths": {
        "/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Projects List",
                "description": "You can get your projects",
                "operationId": "e66a4b9c561c9e3e0b98d95b6437ff91",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of tasks per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 30
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Skip number of tasks for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string",
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "title": "Project 1",
                                        "description": "Project description"
                                    },
                                    {
                                        "id": 43,
                                        "title": "Project 2",
                                        "description": null
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ErrorResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unathorized"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Tasks List",
                "description": "You can get your tasks",
                "operationId": "7786da811325606208249ebc8d393d9a",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of tasks per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 30
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Skip number of tasks for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "project_title": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "priority": {
                                                "type": "string"
                                            },
                                            "due_date": {
                                                "type": "string"
                                            },
                                            "deadline": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": [
                                    {
                                        "id": 1,
                                        "title": "Task 1",
                                        "description": "Task description",
                                        "project_title": "Project 1",
                                        "status": "Incomplete",
                                        "priority": "Medium",
                                        "due_date": "1403/12/11",
                                        "deadline": "1403/12/15"
                                    },
                                    {
                                        "id": 24,
                                        "title": "Task 30",
                                        "description": null,
                                        "project_title": "Project 3",
                                        "status": "Complete",
                                        "priority": "High",
                                        "due_date": "1403/12/22",
                                        "deadline": "1403/12/28"
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ErrorResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unathorized"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/tasks/store": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create Task",
                "description": "You can create task",
                "operationId": "efc4cef55032f4ce69358f0f2e796eb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "title",
                                    "due_date",
                                    "deadline",
                                    "priority"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Task 1"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Task description"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "example": "1403/12/11"
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "example": "1403/12/22"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high"
                                        ],
                                        "example": "low"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "project_title": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "priority": {
                                                "type": "string"
                                            },
                                            "due_date": {
                                                "type": "string"
                                            },
                                            "deadline": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": {
                                    "id": 1,
                                    "title": "Task 1",
                                    "description": "Task description",
                                    "project_title": "Project 1",
                                    "status": "Incomplete",
                                    "priority": "Medium",
                                    "due_date": "1403/12/11",
                                    "deadline": "1403/12/15"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ErrorResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unathorized"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/tasks/{taskId}/complete": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Complete task",
                "description": "Mark task as completed",
                "operationId": "a8f58dcbc205542c5fcb0f9734d8a18e",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID of the task to complete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "message": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": {
                                    "message": "Task updated successfully"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ErrorResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unathorized"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/tasks/{taskId}": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete task",
                "description": "Delete task",
                "operationId": "fd1f7e7ce0772a056265defecf813e67",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID of the task to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "message": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "example": {
                                    "message": "Task deleted successfully"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ErrorResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unathorized"
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "responses": {
            "ErrorResponse": {
                "description": "Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "description": "Show error message to user",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Error message"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "Unathorized": {
                "description": "When username or password is wrong"
            },
            "NotFound": {
                "description": "When item not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "description": "Show error message to user",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Error message"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "Projects"
        },
        {
            "name": "Tasks",
            "description": "Tasks"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ]
}